Добавил дополнительные флаги компилятора в CMAKE и исправил выявленные ошибки. Вынес общие флаги для микроконтроллера и хоста в отдельный файл.

+++ application/tests/mocks/clock-mock.cpp
@@ -8,6 +8,6 @@ namespace ApplicationTests
 		time_t now = time(nullptr);
 		tm* ltm = localtime(&now);
 
-		return ltm->tm_hour * 3600 + ltm->tm_min * 60 + ltm->tm_sec;
+		return static_cast<uint32_t>(ltm->tm_hour * 3600 + ltm->tm_min * 60 + ltm->tm_sec);
 	}
 }
